requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pdf2image==1.17.0
pillow==10.3.0
pypdf==4.2.0
requests==2.32.3
urllib3==2.2.1
Werkzeug==3.0.3
